Changes done in pricing page
